requirements.txt
asgiref==3.4.1
Django==3.2.13
django-filter==21.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
phonenumberslite==8.12.46
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
